+++ internal/sqlparse/sqlparse_test.go
@@ -9,50 +9,183 @@ import (
 
 	"github.com/eugenetriguba/bolt/internal/sqlparse"
 	"github.com/eugenetriguba/checkmate/assert"
+	"github.com/eugenetriguba/checkmate/check"
 )
 
 func TestSqlParser_Parse(t *testing.T) {
 	testCases := []struct {
-		buffer                   string
-		expectedExecutionOptions sqlparse.ExecutionOptions
+		migration               string
+		expectedUpgradeScript   sqlparse.MigrationScript
+		expectedDowngradeScript sqlparse.MigrationScript
 	}{
 		{
-			buffer:                   "-- bolt: no-transaction",
-			expectedExecutionOptions: sqlparse.ExecutionOptions{UseTransaction: false},
+			migration: "",
+			expectedUpgradeScript: sqlparse.MigrationScript{
+				Contents: "",
+				Options:  sqlparse.ExecutionOptions{},
+			},
+			expectedDowngradeScript: sqlparse.MigrationScript{
+				Contents: "",
+				Options:  sqlparse.ExecutionOptions{},
+			},
 		},
 		{
-			buffer:                   "-- BOLT: no-transaction",
-			expectedExecutionOptions: sqlparse.ExecutionOptions{UseTransaction: false},
+			migration: `
+			-- migrate:up
+			CREATE TABLE users(id int PRIMARY KEY);`,
+			expectedUpgradeScript: sqlparse.MigrationScript{
+				Contents: "CREATE TABLE users(id int PRIMARY KEY);\n",
+				Options:  sqlparse.ExecutionOptions{UseTransaction: true},
+			},
+			expectedDowngradeScript: sqlparse.MigrationScript{
+				Contents: "",
+				Options:  sqlparse.ExecutionOptions{},
+			},
 		},
 		{
-			buffer:                   "\n-- bolt: no-transaction",
-			expectedExecutionOptions: sqlparse.ExecutionOptions{UseTransaction: true},
+			migration: `
+			-- migrate:up transaction:false
+			CREATE TABLE users(id int PRIMARY KEY);`,
+			expectedUpgradeScript: sqlparse.MigrationScript{
+				Contents: "CREATE TABLE users(id int PRIMARY KEY);\n",
+				Options:  sqlparse.ExecutionOptions{UseTransaction: false},
+			},
+			expectedDowngradeScript: sqlparse.MigrationScript{
+				Contents: "",
+				Options:  sqlparse.ExecutionOptions{},
+			},
 		},
 		{
-			buffer:                   "",
-			expectedExecutionOptions: sqlparse.ExecutionOptions{UseTransaction: true},
+			migration: `
+			-- migrate:up transaction:true
+			CREATE TABLE users(id int PRIMARY KEY);`,
+			expectedUpgradeScript: sqlparse.MigrationScript{
+				Contents: "CREATE TABLE users(id int PRIMARY KEY);\n",
+				Options:  sqlparse.ExecutionOptions{UseTransaction: true},
+			},
+			expectedDowngradeScript: sqlparse.MigrationScript{
+				Contents: "",
+				Options:  sqlparse.ExecutionOptions{},
+			},
 		},
 		{
-			buffer:                   "--bolt: no-transaction",
-			expectedExecutionOptions: sqlparse.ExecutionOptions{UseTransaction: true},
+			migration: `
+			-- transaction:true migrate:up
+			CREATE TABLE users(id int PRIMARY KEY);`,
+			expectedUpgradeScript: sqlparse.MigrationScript{
+				Contents: "",
+				Options:  sqlparse.ExecutionOptions{},
+			},
+			expectedDowngradeScript: sqlparse.MigrationScript{
+				Contents: "",
+				Options:  sqlparse.ExecutionOptions{},
+			},
 		},
 		{
-			buffer:                   "-- bolt : no-transaction",
-			expectedExecutionOptions: sqlparse.ExecutionOptions{UseTransaction: true},
+			migration: `
+			-- migrate:down
+			DROP TABLE users;`,
+			expectedUpgradeScript: sqlparse.MigrationScript{
+				Contents: "",
+				Options:  sqlparse.ExecutionOptions{},
+			},
+			expectedDowngradeScript: sqlparse.MigrationScript{
+				Contents: "DROP TABLE users;\n",
+				Options:  sqlparse.ExecutionOptions{UseTransaction: true},
+			},
 		},
 		{
-			buffer:                   "-- bolt: no-transaction no-transaction",
-			expectedExecutionOptions: sqlparse.ExecutionOptions{UseTransaction: false},
+			migration: `
+			-- migrate:down transaction:false
+			DROP TABLE users;`,
+			expectedUpgradeScript: sqlparse.MigrationScript{
+				Contents: "",
+				Options:  sqlparse.ExecutionOptions{},
+			},
+			expectedDowngradeScript: sqlparse.MigrationScript{
+				Contents: "DROP TABLE users;\n",
+				Options:  sqlparse.ExecutionOptions{UseTransaction: false},
+			},
+		},
+		{
+			migration: `
+			-- migrate:down transaction:true
+			DROP TABLE users;`,
+			expectedUpgradeScript: sqlparse.MigrationScript{
+				Contents: "",
+				Options:  sqlparse.ExecutionOptions{},
+			},
+			expectedDowngradeScript: sqlparse.MigrationScript{
+				Contents: "DROP TABLE users;\n",
+				Options:  sqlparse.ExecutionOptions{UseTransaction: true},
+			},
+		},
+		{
+			migration: `
+			-- transaction:true migrate:down
+			DROP TABLE users;`,
+			expectedUpgradeScript: sqlparse.MigrationScript{
+				Contents: "",
+				Options:  sqlparse.ExecutionOptions{},
+			},
+			expectedDowngradeScript: sqlparse.MigrationScript{
+				Contents: "",
+				Options:  sqlparse.ExecutionOptions{},
+			},
+		},
+		{
+			migration: `
+			-- migrate:up
+			CREATE TABLE users(id int PRIMARY KEY);
+			-- migrate:down
+			DROP TABLE users;`,
+			expectedUpgradeScript: sqlparse.MigrationScript{
+				Contents: "CREATE TABLE users(id int PRIMARY KEY);\n",
+				Options:  sqlparse.ExecutionOptions{UseTransaction: true},
+			},
+			expectedDowngradeScript: sqlparse.MigrationScript{
+				Contents: "DROP TABLE users;\n",
+				Options:  sqlparse.ExecutionOptions{UseTransaction: true},
+			},
+		},
+		{
+			migration: `
+			-- migrate:up transaction:false
+			CREATE TABLE users(id int PRIMARY KEY);
+			-- migrate:down transaction:false
+			DROP TABLE users;`,
+			expectedUpgradeScript: sqlparse.MigrationScript{
+				Contents: "CREATE TABLE users(id int PRIMARY KEY);\n",
+				Options:  sqlparse.ExecutionOptions{UseTransaction: false},
+			},
+			expectedDowngradeScript: sqlparse.MigrationScript{
+				Contents: "DROP TABLE users;\n",
+				Options:  sqlparse.ExecutionOptions{UseTransaction: false},
+			},
+		},
+		{
+			migration: `
+			-- MIGRATE:UP TRANSACTION:FALSE
+			CREATE TABLE users(id int PRIMARY KEY);
+			-- MIGRATE:DOWN TRANSACTION:FALSE
+			DROP TABLE users;`,
+			expectedUpgradeScript: sqlparse.MigrationScript{
+				Contents: "CREATE TABLE users(id int PRIMARY KEY);\n",
+				Options:  sqlparse.ExecutionOptions{UseTransaction: false},
+			},
+			expectedDowngradeScript: sqlparse.MigrationScript{
+				Contents: "DROP TABLE users;\n",
+				Options:  sqlparse.ExecutionOptions{UseTransaction: false},
+			},
 		},
 	}
 
 	for _, tc := range testCases {
 		sqlParser := sqlparse.NewSqlParser()
-
-		execOptions, err := sqlParser.Parse(strings.NewReader(tc.buffer))
+		upgradeScript, downgradeScript, err := sqlParser.Parse(strings.NewReader(tc.migration))
 		assert.Nil(t, err)
-
-		assert.DeepEqual(t, execOptions, tc.expectedExecutionOptions)
+		check.DeepEqual(t, upgradeScript, tc.expectedUpgradeScript)
+		check.DeepEqual(t, downgradeScript, tc.expectedDowngradeScript)
 	}
 }
 
@@ -80,7 +213,7 @@ func TestSqlParserParseError(t *testing.T) {
 		ErrCond: " ",
 	}
 	sqlParser := sqlparse.NewSqlParser()
-	_, err := sqlParser.Parse(reader)
+	_, _, err := sqlParser.Parse(reader)
 	assert.NotNil(t, err)
 	assert.ErrorContains(t, err, "unwanted input encountered")
 }